move Mathematica comparison tests
